requirements.txt
streamlit
openai
python-dotenv
pyyaml
pytz
arabic-reshaper
python-bidi
streamlit-option-menu